refactor/#19 : 소통 관련 엔티티 수정(코드 리뷰, 현직자 조언 구분)
