+++ src/main/java/com/example/devcrew/domain/feedback/dto/request/CreateCodeFeedbackRequestDTO.java
@@ -1,7 +1,6 @@
-package com.example.devcrew.domain.feedback.dto;
+package com.example.devcrew.domain.feedback.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 
 @Getter
@@ -13,7 +12,6 @@ public class CreateCodeFeedbackRequestDTO {
     @NotBlank
     private String content; // 코드 리뷰 게시글 내용
 
-    @NotNull
     private Integer language;    // 태그를 번호로 받을 것
 
     private String fileUrl;  // 첨부 파일 리스트
