+++ src/main/java/com/example/devcrew/domain/feedback/entity/AdviceFeedback.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class AdviceFeedback{
 
     @Enumerated(EnumType.STRING)
-    @Column(name = "feedback_tag", nullable = false, length = 20)
+    @Column(name = "feedback_tag", nullable = true, length = 20)
     private FeedbackTag feedbackTag;  // 게시글 관련 태그(기획, 디자인, FE, BE, 기타)
 
     @Builder
