+++ src/main/java/com/example/devcrew/domain/feedback/repository/FeedbackRepository.java
@@ -1,6 +1,7 @@
 package com.example.devcrew.domain.feedback.repository;
 
 import com.example.devcrew.domain.feedback.entity.Feedback;
+import com.example.devcrew.domain.feedback.entity.FeedbackTag;
 import com.example.devcrew.domain.member.entity.Member;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -8,5 +9,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FeedbackRepository extends JpaRepository<Feedback, Long> {
 
-    Page<Feedback> findByMember(Member member, PageRequest pageRequest);
+    Page<Feedback> findByAdviceFeedback_FeedbackTag(FeedbackTag feedbackTag, PageRequest pageRequest);
 }
