+++ src/main/java/com/example/devcrew/domain/feedback/service/CreateAdviceFeedback.java
@@ -1,9 +1,7 @@
 package com.example.devcrew.domain.feedback.service;
 
-import com.example.devcrew.domain.feedback.dto.CreateAdviceFeedbackRequestDTO;
-import com.example.devcrew.domain.feedback.entity.AdviceFeedback;
+import com.example.devcrew.domain.feedback.dto.request.CreateAdviceFeedbackRequestDTO;
 import com.example.devcrew.domain.feedback.entity.Feedback;
-import org.springframework.security.core.userdetails.UserDetails;
 
 public interface CreateAdviceFeedback {
