+++ src/main/java/com/example/devcrew/domain/feedback/service/CreateCodeFeedback.java
@@ -1,6 +1,6 @@
 package com.example.devcrew.domain.feedback.service;
 
-import com.example.devcrew.domain.feedback.dto.CreateCodeFeedbackRequestDTO;
+import com.example.devcrew.domain.feedback.dto.request.CreateCodeFeedbackRequestDTO;
 import com.example.devcrew.domain.feedback.entity.Feedback;
 
 public interface CreateCodeFeedback {
