+++ src/main/java/com/example/devcrew/domain/feedback/service/ReadAdviceFeedback.java
@@ -1,6 +1,6 @@
 package com.example.devcrew.domain.feedback.service;
 
-import com.example.devcrew.domain.feedback.dto.ReadAdviceFeedbackResponseDTO;
+import com.example.devcrew.domain.feedback.dto.response.ReadAdviceFeedbackResponseDTO;
 
 public interface ReadAdviceFeedback {
     ReadAdviceFeedbackResponseDTO readAdviceFeedback(Long feedbackId);
