+++ src/main/java/com/example/devcrew/domain/feedback/service/ReadCodeFeedback.java
@@ -1,7 +1,6 @@
 package com.example.devcrew.domain.feedback.service;
 
-import com.example.devcrew.domain.feedback.dto.ReadAdviceFeedbackResponseDTO;
-import com.example.devcrew.domain.feedback.dto.ReadCodeFeedbackResponseDTO;
+import com.example.devcrew.domain.feedback.dto.response.ReadCodeFeedbackResponseDTO;
 
 public interface ReadCodeFeedback {
     ReadCodeFeedbackResponseDTO readCodeFeedback(Long feedbackId);
